# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Directory Master API
servers:
  - url: /api
    description: Base API path
...
paths:
  /contacts:
    post:
      operationId: createContact
      tags: [contacts]
      summary: Submit a claim listing form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSubmitBody"
      responses:
        "201":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contact submissions (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
  /contacts/{id}:
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact submission (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List all entries (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: published
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryListResponse"

    post:
      operationId: createEntry
      tags: [entries]
      summary: Create a new entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntryBody"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /entries/{id}:
    get:
      operationId: getEntry
      tags: [entries]
      summary: Get a single entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateEntry
      tags: [entries]
      summary: Update an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntryBody"
      responses:
        "200":
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"

    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Delete an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /entries/{id}/publish:
    patch:
      operationId: toggleEntryPublished
      tags: [entries]
      summary: Toggle entry published status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TogglePublishBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /public/entries:
    get:
      operationId: listPublicEntries
      tags: [public]
      summary: List published entries for public directory
      parameters:
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: sort
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Published entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryListResponse"
  /public/entries/{id}:
    get:
      operationId: getPublicEntry
      tags: [public]
      summary: Get a single published entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required: [success]
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, editor, viewer]
        createdAt:
          type: string
      required: [id, name, email, role, createdAt]
...
    Entry:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        eventType:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        moreDetails:
          type: ["string", "null"]
        customFields:
          type: ["object", "null"]
        sourceCsvRow:
          type: ["string", "null"]
        published:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, published, createdAt, updatedAt]
    CreateEntryBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        eventType:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        moreDetails:
          type: ["string", "null"]
        published:
          type: ["boolean", "null"]
      required: [title]
    UpdateEntryBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        eventType:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        moreDetails:
          type: ["string", "null"]
        customFields:
          type: ["object", "null"]
        published:
          type: ["boolean", "null"]
    TogglePublishBody:
      type: object
      properties:
        published:
          type: boolean
      required: [published]
    EntryListResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [entries, total, page, totalPages]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        entryCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, slug, entryCount, createdAt]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required: [name]
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
...
    ContactSubmitBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
      required: [fullName, phone, email]
    Contact:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required: [id, fullName, phone, email, createdAt]
    ContactListResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
      required: [contacts]
...