# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directory Master API
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}